requirements.txt
Django==3.0.7
APSCheduler==3.6.3
django-tempus-dominus==5.1.2.13
django-tinymce==3.0.2
channels==2.4.0
requests==2.23.0
Pillow==7.1.2
Pyrogram==0.17.1
TgCrypto==1.2.1